refactor: srs.py を store.py に改名し責務を明確化（互換エイリアス追加、参照更新、不要ファイル削除、ドキュメント同期）
