+++ UserManual.md
@@ -84,7 +84,7 @@
 
 ### A-11. 制約・既知の事項（ユーザー向け）
 - UI文言は今後統一予定
-- 復習（SRS）は自動で保存されます（初回起動時に数枚のカードが用意されます）
+- 復習（SRS）は同じSQLiteに保存されます（初回起動時に数枚のカードが用意されます）
 
 ### A-12. トラブルシュート（一般ユーザー向け）
 - 表示が更新されない: ブラウザを更新、または一度アプリを停止して再起動
