--- app/srs.py
@@ -1,9 +1,0 @@
-class SpacedRepetitionScheduler:
-    """Placeholder for spaced repetition system (SRS) logic."""
-
-    def schedule(self, item: str, performance: int) -> None:
-        """Schedule the next review for an item based on performance.
-
-        This is a placeholder and does not implement any real scheduling logic.
-        """
-        raise NotImplementedError("SRS algorithm not implemented yet")
