+++ src/backend/flows/article_import.py
@@ -7,7 +7,7 @@ import uuid
 from ..config import settings
 from ..logging import logger
 from ..providers import get_llm_provider
-from ..srs import store
+from ..store import store
 from ..models.word import WordPack
 from ..models.article import (
     ArticleImportRequest,
