+++ 終了済みor参考ドキュメント/環境変数の意味.md
@@ -163,9 +163,9 @@ _rag_executor = ThreadPoolExecutor(max_workers=4)
   - `SRS_MAX_TODAY`: 1日の出題上限（将来用のパラメータ）
 - 既定値: env.example は `.data/srs.sqlite3` と `5`、コード既定も同じ
 - 使われ方:
-```400:402:src/backend/srs.py
+```898:903:src/backend/store.py
 # module-level singleton store (wired to settings)
-store = SRSSQLiteStore(db_path=settings.srs_db_path)
+store = AppSQLiteStore(db_path=settings.srs_db_path)
 ```
   - DB パスは起動時にディレクトリ作成と DB 初期化。  
   - 現状、`SRS_MAX_TODAY` はコード内で参照されていません（未配線）。ルーターは固定値 5 を返します。
